Fixed some implicit conversions

+++ src/Thunderegg/Poisson/StencilHelper.h
@@ -296,7 +296,7 @@ class CoarseSH : public StencilHelper
 		double            h        = 0;
 		int               idx      = d.global_index * d.ns[0] * d.ns[1] * d.ns[2];
 		int               nbr_idx  = nbr_info.global_index * d.ns[0] * d.ns[1] * d.ns[2];
-		quad                       = nbr_info.orth_on_coarse.getIndex();
+		quad                       = (int) nbr_info.orth_on_coarse.getIndex();
 		if (s == Side<3>::west()) {
 			h         = d.spacings[0];
 			start     = idx;
@@ -354,7 +354,7 @@ class CoarseSH : public StencilHelper
 	}
 	int size(int xi, int yi)
 	{
-		return coeff.size();
+		return (int) coeff.size();
 	}
 	double *coeffs(int xi, int yi)
 	{
@@ -487,7 +487,7 @@ class FineSH : public StencilHelper
 	}
 	int size(int xi, int yi)
 	{
-		return coeff.size();
+		return (int) coeff.size();
 	}
 	double *coeffs(int xi, int yi)
 	{
